requirements.txt
numpy>=1.26,<3.0
pandas>=2.2
scikit-learn>=1.4
lightgbm>=4.1
torch>=2.2
matplotlib>=3.8
python-dotenv>=1.0
